+++ MIDAS/src/gnc/yessir.cpp
@@ -272,30 +272,30 @@ void Yessir::setState(KalmanState state) {
  * updated based on the time taken per cycle of the Kalman Filter Thread.
  */
 void Yessir::setQ(float dt, float sd) {
-    Q(0, 0) = pow(dt, 5) / 20;
-    Q(0, 1) = pow(dt, 4) / 8;
-    Q(0, 2) = pow(dt, 3) / 6;
-    Q(1, 1) = pow(dt, 3) / 8;
-    Q(1, 2) = pow(dt, 2) / 2;
+    Q(0, 0) = powf(dt, 5) / 20;
+    Q(0, 1) = powf(dt, 4) / 8;
+    Q(0, 2) = powf(dt, 3) / 6;
+    Q(1, 1) = powf(dt, 3) / 8;
+    Q(1, 2) = powf(dt, 2) / 2;
     Q(2, 2) = dt;
     Q(1, 0) = Q(0, 1);
     Q(2, 0) = Q(0, 2);
     Q(2, 1) = Q(1, 2);
-    Q(3, 3) = pow(dt, 5) / 20;
-    Q(3, 4) = pow(dt, 4) / 8;
-    Q(3, 5) = pow(dt, 3) / 6;
-    Q(4, 4) = pow(dt, 3) / 8;
-    Q(4, 5) = pow(dt, 2) / 2;
+    Q(3, 3) = powf(dt, 5) / 20;
+    Q(3, 4) = powf(dt, 4) / 8;
+    Q(3, 5) = powf(dt, 3) / 6;
+    Q(4, 4) = powf(dt, 3) / 8;
+    Q(4, 5) = powf(dt, 2) / 2;
     Q(5, 5) = dt;
     Q(4, 3) = Q(3, 4);
     Q(5, 3) = Q(3, 5);
     Q(5, 4) = Q(4, 5);
 
-    Q(6, 6) = pow(dt, 5) / 20;
-    Q(6, 7) = pow(dt, 4) / 8;
-    Q(6, 8) = pow(dt, 3) / 6;
-    Q(7, 7) = pow(dt, 3) / 8;
-    Q(7, 8) = pow(dt, 2) / 2;
+    Q(6, 6) = powf(dt, 5) / 20;
+    Q(6, 7) = powf(dt, 4) / 8;
+    Q(6, 8) = powf(dt, 3) / 6;
+    Q(7, 7) = powf(dt, 3) / 8;
+    Q(7, 8) = powf(dt, 2) / 2;
     Q(8, 8) = dt;
     Q(7, 6) = Q(6, 7);
     Q(8, 6) = Q(6, 8);
