+++ MIDAS/src/hardware/Voltage.cpp
@@ -4,7 +4,7 @@
 
 #include "pins.h"
 
-#define VOLTAGE_DIVIDER (5.0 / (5.0 + 20.0))
+#define VOLTAGE_DIVIDER (5.0f / (5.0f + 20.0f))
 
 /**
  * @brief "Initializes" the voltage sensor. Since it reads directly from a pin without a library, there is no specific initialization.
@@ -22,7 +22,7 @@ ErrorCode VoltageSensor::init() {
 */
 VoltageData VoltageSensor::read() {
     VoltageData v_battery;
-    v_battery.voltage = adcAnalogRead(ADCAddress{VOLTAGE_PIN}).value * 3.3f / 4095.0f / VOLTAGE_DIVIDER;
+    v_battery.voltage = static_cast<float>(adcAnalogRead(ADCAddress{VOLTAGE_PIN}).value) * 3.3f / 4095.0f / VOLTAGE_DIVIDER;
 //    Serial.print("Raw voltage reading: ");
 //    Serial.print(v_battery.voltage);
 //    Serial.println("");
